+++ src/main/java/org/snomed/snowstormlite/domain/FHIRMapping.java
@@ -8,12 +8,14 @@ public class FHIRMapping {
 	private String code;
 	private String correlation;
 	private String message;
+	private boolean inverse;
 
-	public FHIRMapping(String refsetId, String code, String correlation, String message) {
+	public FHIRMapping(String refsetId, String code, String correlation, String message, boolean inverse) {
 		this.refsetId = refsetId;
 		this.code = code;
 		this.correlation = correlation;
 		this.message = message;
+		this.inverse = inverse;
 	}
 
 	public String getRefsetId() {
@@ -48,13 +50,18 @@ public class FHIRMapping {
 		this.message = message;
 	}
 
+	public boolean isInverse() {
+		return inverse;
+	}
+
 	public String toIndexString() {
-		return format("%s|%s|%s|%s", refsetId, stringToIndex(code), stringToIndex(correlation), stringToIndex(message));
+		return format("%s|%s|%s|%s|%s", refsetId, stringToIndex(code), stringToIndex(correlation), stringToIndex(message), inverse ? "1" : "0");
 	}
 
 	public static FHIRMapping fromIndexString(String indexString) {
 		String[] split = indexString.split("\\|");
-		return new FHIRMapping(split[0], getStringFromIndex(split, 1), getStringFromIndex(split, 2), getStringFromIndex(split, 3));
+		return new FHIRMapping(split[0], getStringFromIndex(split, 1), getStringFromIndex(split, 2), getStringFromIndex(split, 3),
+				"1".equals(split[4]));
 	}
 
 	private String stringToIndex(String value) {